requirements.txt
Django==5.0.1
Jinja2==3.1.3
django-jinja==2.11.0
psycopg2-binary==2.9.9
Pillow==10.2.0
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
django-extensions==3.2.3
celery==5.3.4
redis==5.0.1
django-redis==5.4.0
stripe==7.8.0
razorpay==1.4.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
djangorestframework==3.14.0
django-allauth==0.57.0

